refactor: create component for download menu
